requirements.txt
Django==3.0.6
channels==2.4.0
pillow==7.1.2
opencv-python==4.2.0.34
numpy==1.18.5
redis==6.0.4